move window helper into helper directory
