+++ WinUIGallery/ControlPages/SplitViewPage.xaml.cs
@@ -1,3 +1,4 @@
+using AppUIBasics.Helper;
 using System;
 using System.Collections.Generic;
 using System.IO;
