+++ WinUIGallery/IdleSynchronizer.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
+using AppUIBasics.Helper;
 using System;
 using System.Runtime.InteropServices;
 using System.Threading;
